Separate code from fork configs
